Reorganize code.
Move XmlInjectable under Element namespace in order
to avoid name conflicts.
