+++ src/Element/Service.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Raigu\XRoad\SoapEnvelope;
+namespace Raigu\XRoad\SoapEnvelope\Element;
 
 use DOMDocument;
 
 /**
- * I am a client part of X-Road SOAP Envelope
+ * I am a service part of X-Road SOAP Envelope
  */
-class Client implements XmlInjectable
+class Service implements XmlInjectable
 {
     /**
      * @var array
@@ -16,23 +16,22 @@ class Client implements XmlInjectable
 
     public function inject(DOMDocument $dom)
     {
-        $client = $dom->createElementNS(
+        $service = $dom->createElementNS(
             'http://x-road.eu/xsd/xroad.xsd',
-            'client'
+            'service'
         );
 
         $objectType = $dom->createAttributeNS(
             'http://x-road.eu/xsd/identifiers',
             'objectType'
         );
-        $objectType->value = 'SUBSYSTEM';
-        $client->appendChild($objectType);
+        $objectType->value = 'SERVICE';
+        $service->appendChild($objectType);
 
-        $names = ['xRoadInstance', 'memberClass', 'memberCode', 'subsystemCode'];
-
+        $names = ['xRoadInstance', 'memberClass', 'memberCode', 'subsystemCode', 'serviceCode', 'serviceVersion'];
 
         foreach (array_combine($names, $this->values) as $name => $value) {
-            $client->appendChild(
+            $service->appendChild(
                 $dom->createElementNS(
                     'http://x-road.eu/xsd/identifiers',
                     $name,
@@ -46,19 +45,19 @@ class Client implements XmlInjectable
             'Header'
         );
 
-        $elements->item(0)->appendChild($client);
+        $elements->item(0)->appendChild($service);
     }
 
     public static function fromStr(string $value): self
     {
         $values = explode('/', $value);
-        if (count($values) !== 4) {
-            throw new \Exception('Could not extract client parameters. Invalid format.');
+        if (count($values) !== 6) {
+            throw new \Exception('Could not extract service parameters. Invalid format.');
         }
         return new self($values);
     }
 
-    protected function __construct(array $values)
+    private function __construct(array $values)
     {
         $this->values = $values;
     }
